+++ x-pack/plugins/apm/server/lib/alerts/register_transaction_error_rate_alert_type.ts
@@ -5,25 +5,25 @@
  */
 
 import { schema } from '@kbn/config-schema';
+import { isEmpty } from 'lodash';
 import { Observable } from 'rxjs';
 import { take } from 'rxjs/operators';
-import { isEmpty } from 'lodash';
-import { asDecimalOrInteger } from '../../../common/utils/formatters';
-import { ProcessorEvent } from '../../../common/processor_event';
-import { EventOutcome } from '../../../common/event_outcome';
+import { APMConfig } from '../..';
+import { ESSearchResponse } from '../../../../../typings/elasticsearch';
+import { AlertingPlugin } from '../../../../alerts/server';
 import { AlertType, ALERT_TYPES_CONFIG } from '../../../common/alert_types';
-import { ESSearchResponse } from '../../../typings/elasticsearch';
 import {
+  EVENT_OUTCOME,
   PROCESSOR_EVENT,
+  SERVICE_ENVIRONMENT,
   SERVICE_NAME,
   TRANSACTION_TYPE,
-  EVENT_OUTCOME,
-  SERVICE_ENVIRONMENT,
 } from '../../../common/elasticsearch_fieldnames';
-import { AlertingPlugin } from '../../../../alerts/server';
-import { getApmIndices } from '../settings/apm_indices/get_apm_indices';
-import { APMConfig } from '../..';
+import { EventOutcome } from '../../../common/event_outcome';
+import { ProcessorEvent } from '../../../common/processor_event';
+import { asDecimalOrInteger } from '../../../common/utils/formatters';
 import { getEnvironmentUiFilterES } from '../helpers/convert_ui_filters/get_environment_ui_filter_es';
+import { getApmIndices } from '../settings/apm_indices/get_apm_indices';
 import { apmActionVariables } from './action_variables';
 
 interface RegisterAlertParams {
