+++ x-pack/plugins/apm/server/lib/helpers/get_internal_saved_objects_client.ts
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import { CoreSetup } from 'src/core/server';
-import { PromiseReturnType } from '../../../typings/common';
+import { PromiseReturnType } from '../../../../observability/typings/common';
 
 export type InternalSavedObjectsClient = PromiseReturnType<
   typeof getInternalSavedObjectsClient
