+++ x-pack/plugins/apm/server/lib/helpers/setup_request.ts
@@ -4,20 +4,21 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import moment from 'moment';
 import { Logger } from 'kibana/server';
-import { isActivePlatinumLicense } from '../../../common/service_map';
-import { UI_SETTINGS } from '../../../../../../src/plugins/data/common';
-import { KibanaRequest } from '../../../../../../src/core/server';
+import moment from 'moment';
 import { APMConfig } from '../..';
-import {
-  getApmIndices,
-  ApmIndicesConfig,
-} from '../settings/apm_indices/get_apm_indices';
-import { ESFilter } from '../../../typings/elasticsearch';
-import { getEsFilter } from './convert_ui_filters/get_es_filter';
-import { APMRequestHandlerContext } from '../../routes/typings';
+import { KibanaRequest } from '../../../../../../src/core/server';
+import { UI_SETTINGS } from '../../../../../../src/plugins/data/common';
+import { ESFilter } from '../../../../../typings/elasticsearch';
 import { ProcessorEvent } from '../../../common/processor_event';
+import { isActivePlatinumLicense } from '../../../common/service_map';
+import { UIFilters } from '../../../typings/ui_filters';
+import { APMRequestHandlerContext } from '../../routes/typings';
+import {
+  ApmIndicesConfig,
+  getApmIndices,
+} from '../settings/apm_indices/get_apm_indices';
+import { getEsFilter } from './convert_ui_filters/get_es_filter';
 import {
   APMEventClient,
   createApmEventClient,
@@ -26,7 +27,6 @@ import {
   APMInternalClient,
   createInternalESClient,
 } from './create_es_client/create_internal_es_client';
-import { UIFilters } from '../../../typings/ui_filters';
 
 // Explicitly type Setup to prevent TS initialization errors
 // https://github.com/microsoft/TypeScript/issues/34933
