+++ x-pack/plugins/apm/server/lib/service_map/get_service_map_service_node_info.ts
@@ -4,33 +4,33 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { ESFilter } from '../../../../../typings/elasticsearch';
 import {
-  TRANSACTION_REQUEST,
-  TRANSACTION_PAGE_LOAD,
-} from '../../../common/transaction_types';
-import {
-  SERVICE_NAME,
+  METRIC_CGROUP_MEMORY_USAGE_BYTES,
   METRIC_SYSTEM_CPU_PERCENT,
   METRIC_SYSTEM_FREE_MEMORY,
   METRIC_SYSTEM_TOTAL_MEMORY,
-  METRIC_CGROUP_MEMORY_USAGE_BYTES,
+  SERVICE_NAME,
   TRANSACTION_TYPE,
 } from '../../../common/elasticsearch_fieldnames';
 import { ProcessorEvent } from '../../../common/processor_event';
+import {
+  TRANSACTION_PAGE_LOAD,
+  TRANSACTION_REQUEST,
+} from '../../../common/transaction_types';
 import { rangeFilter } from '../../../common/utils/range_filter';
-import { ESFilter } from '../../../typings/elasticsearch';
+import {
+  getDocumentTypeFilterForAggregatedTransactions,
+  getProcessorEventForAggregatedTransactions,
+  getTransactionDurationFieldForAggregatedTransactions,
+} from '../helpers/aggregated_transactions';
+import { getEnvironmentUiFilterES } from '../helpers/convert_ui_filters/get_environment_ui_filter_es';
 import { Setup, SetupTimeRange } from '../helpers/setup_request';
 import {
   percentCgroupMemoryUsedScript,
   percentSystemMemoryUsedScript,
 } from '../metrics/by_agent/shared/memory';
-import {
-  getProcessorEventForAggregatedTransactions,
-  getTransactionDurationFieldForAggregatedTransactions,
-  getDocumentTypeFilterForAggregatedTransactions,
-} from '../helpers/aggregated_transactions';
 import { getErrorRate } from '../transaction_groups/get_error_rate';
-import { getEnvironmentUiFilterES } from '../helpers/convert_ui_filters/get_environment_ui_filter_es';
 
 interface Options {
   setup: Setup & SetupTimeRange;
