+++ x-pack/plugins/apm/server/lib/settings/agent_configuration/find_exact_configuration.ts
@@ -4,13 +4,13 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { ESSearchHit } from '../../../../../../typings/elasticsearch';
+import { AgentConfiguration } from '../../../../common/agent_configuration/configuration_types';
 import {
-  SERVICE_NAME,
   SERVICE_ENVIRONMENT,
+  SERVICE_NAME,
 } from '../../../../common/elasticsearch_fieldnames';
 import { Setup } from '../../helpers/setup_request';
-import { AgentConfiguration } from '../../../../common/agent_configuration/configuration_types';
-import { ESSearchHit } from '../../../../typings/elasticsearch';
 import { convertConfigSettingsToString } from './convert_settings_to_string';
 
 export async function findExactConfiguration({
