+++ x-pack/plugins/apm/server/lib/transaction_groups/get_transaction_group_stats.ts
@@ -6,7 +6,7 @@
 import { merge } from 'lodash';
 import { TRANSACTION_TYPE } from '../../../common/elasticsearch_fieldnames';
 import { arrayUnionToCallable } from '../../../common/utils/array_union_to_callable';
-import { AggregationInputMap } from '../../../typings/elasticsearch/aggregations';
+import { AggregationInputMap } from '../../../../../typings/elasticsearch';
 import { TransactionGroupRequestBase, TransactionGroupSetup } from './fetcher';
 import { getTransactionDurationFieldForAggregatedTransactions } from '../helpers/aggregated_transactions';
