+++ x-pack/plugins/apm/server/lib/transactions/charts/get_timeseries_data/fetcher.ts
@@ -4,22 +4,22 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { ESFilter } from '../../../../../typings/elasticsearch';
+import { ESFilter } from '../../../../../../../typings/elasticsearch';
+import { PromiseReturnType } from '../../../../../../observability/typings/common';
 import {
   SERVICE_NAME,
   TRANSACTION_NAME,
   TRANSACTION_RESULT,
   TRANSACTION_TYPE,
 } from '../../../../../common/elasticsearch_fieldnames';
-import { PromiseReturnType } from '../../../../../../observability/typings/common';
-import { getBucketSize } from '../../../helpers/get_bucket_size';
 import { rangeFilter } from '../../../../../common/utils/range_filter';
-import { Setup, SetupTimeRange } from '../../../helpers/setup_request';
 import {
+  getDocumentTypeFilterForAggregatedTransactions,
   getProcessorEventForAggregatedTransactions,
   getTransactionDurationFieldForAggregatedTransactions,
-  getDocumentTypeFilterForAggregatedTransactions,
 } from '../../../helpers/aggregated_transactions';
+import { getBucketSize } from '../../../helpers/get_bucket_size';
+import { Setup, SetupTimeRange } from '../../../helpers/setup_request';
 
 export type ESResponse = PromiseReturnType<typeof timeseriesFetcher>;
 export function timeseriesFetcher({
