+++ x-pack/plugins/apm/server/projections/util/merge_projection/index.ts
@@ -3,12 +3,14 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-import { mergeWith, isPlainObject, cloneDeep } from 'lodash';
+import { cloneDeep, isPlainObject, mergeWith } from 'lodash';
 import { DeepPartial } from 'utility-types';
-import { AggregationInputMap } from '../../../../typings/elasticsearch/aggregations';
-import { ESSearchBody } from '../../../../typings/elasticsearch';
-import { Projection } from '../../typings';
+import {
+  AggregationInputMap,
+  ESSearchBody,
+} from '../../../../../../typings/elasticsearch';
 import { APMEventESSearchRequest } from '../../../lib/helpers/create_es_client/create_apm_event_client';
+import { Projection } from '../../typings';
 
 type PlainObject = Record<string | number | symbol, any>;
