+++ x-pack/plugins/lens/server/routes/field_stats.ts
@@ -9,7 +9,7 @@ import DateMath from '@elastic/datemath';
 import { schema } from '@kbn/config-schema';
 import { CoreSetup } from 'src/core/server';
 import { IFieldType } from 'src/plugins/data/common';
-import { ESSearchResponse } from '../../../apm/typings/elasticsearch';
+import { ESSearchResponse } from '../../../../typings/elasticsearch';
 import { FieldStatsResponse, BASE_API_URL } from '../../common';
 
 const SHARD_SIZE = 5000;
