+++ x-pack/test/apm_api_integration/basic/tests/correlations/ranges.ts
@@ -6,10 +6,10 @@
 
 import expect from '@kbn/expect';
 import { format } from 'url';
-import { expectSnapshot } from '../../../common/match_snapshot';
-import { PromiseReturnType } from '../../../../../plugins/apm/typings/common';
-import { FtrProviderContext } from '../../../common/ftr_provider_context';
+import { PromiseReturnType } from '../../../../../plugins/observability/typings/common';
 import archives_metadata from '../../../common/archives_metadata';
+import { FtrProviderContext } from '../../../common/ftr_provider_context';
+import { expectSnapshot } from '../../../common/match_snapshot';
 
 export default function ApiTest({ getService }: FtrProviderContext) {
   const supertest = getService('supertest');
