+++ x-pack/test/apm_api_integration/basic/tests/correlations/slow_durations.ts
@@ -6,10 +6,10 @@
 
 import expect from '@kbn/expect';
 import { format } from 'url';
-import { PromiseReturnType } from '../../../../../plugins/apm/typings/common';
+import { PromiseReturnType } from '../../../../../plugins/observability/typings/common';
+import archives_metadata from '../../../common/archives_metadata';
 import { FtrProviderContext } from '../../../common/ftr_provider_context';
 import { expectSnapshot } from '../../../common/match_snapshot';
-import archives_metadata from '../../../common/archives_metadata';
 
 export default function ApiTest({ getService }: FtrProviderContext) {
   const supertest = getService('supertest');
