+++ x-pack/test/apm_api_integration/basic/tests/transaction_groups/transaction_charts.ts
@@ -5,7 +5,7 @@
  */
 import expect from '@kbn/expect';
 import archives_metadata from '../../../common/archives_metadata';
-import { PromiseReturnType } from '../../../../../plugins/apm/typings/common';
+import { PromiseReturnType } from '../../../../../plugins/observability/typings/common';
 import { expectSnapshot } from '../../../common/match_snapshot';
 import { FtrProviderContext } from '../../../../common/ftr_provider_context';
