+++ x-pack/test/apm_api_integration/common/authentication.ts
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { PromiseReturnType } from '../../../plugins/apm/typings/common';
+import { PromiseReturnType } from '../../../plugins/observability/typings/common';
 import { SecurityServiceProvider } from '../../../../test/common/services/security';
 
 type SecurityService = PromiseReturnType<typeof SecurityServiceProvider>;
