+++ x-pack/test/apm_api_integration/trial/tests/services/transaction_groups_charts.ts
@@ -6,7 +6,7 @@
 
 import expect from '@kbn/expect';
 import { expectSnapshot } from '../../../common/match_snapshot';
-import { PromiseReturnType } from '../../../../../plugins/apm/typings/common';
+import { PromiseReturnType } from '../../../../../plugins/observability/typings/common';
 import { FtrProviderContext } from '../../../common/ftr_provider_context';
 import archives_metadata from '../../../common/archives_metadata';
