+++ x-pack/typings/elasticsearch/index.ts
@@ -6,6 +6,17 @@
 
 import { SearchParams, SearchResponse } from 'elasticsearch';
 import { AggregationResponseMap, AggregationInputMap } from './aggregations';
+export {
+  AggregationInputMap,
+  AggregationOptionsByType,
+  AggregationResponseMap,
+  AggregationResultOf,
+  SortOptions,
+  ValidAggregationKeysOf,
+} from './aggregations';
+
+// Typings for Elasticsearch queries and aggregations. These are intended to be
+// moved to the Elasticsearch JS client at some point (see #77720.)
 
 interface CollapseQuery {
   field: string;
@@ -42,10 +53,7 @@ export type ESSearchResponse<
 > = Omit<SearchResponse<TDocument>, 'aggregations' | 'hits'> &
   (TSearchRequest extends { body: { aggs: AggregationInputMap } }
     ? {
-        aggregations?: AggregationResponseMap<
-          TSearchRequest['body']['aggs'],
-          TDocument
-        >;
+        aggregations?: AggregationResponseMap<TSearchRequest['body']['aggs'], TDocument>;
       }
     : {}) & {
     hits: Omit<SearchResponse<TDocument>['hits'], 'total'> &
@@ -63,12 +71,6 @@ export type ESSearchResponse<
 
 export interface ESFilter {
   [key: string]: {
-    [key: string]:
-      | string
-      | string[]
-      | number
-      | boolean
-      | Record<string, unknown>
-      | ESFilter[];
+    [key: string]: string | string[] | number | boolean | Record<string, unknown> | ESFilter[];
   };
 }
